requirements.txt
torchvision==0.8.1
matplotlib==3.3.3
torch==1.7.0
numpy==1.18.5
pandas==1.1.4
Click==7.0
tqdm==4.54.1
opencv_python==4.4.0.46
fire==0.3.1
Pillow==8.0.1
scikit_learn==0.23.2
tensorboardX==2.1
